requirements.txt
absl-py==1.4.0
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
blinker==1.6.3
breadability==0.1.20
cachetools==5.3.1
catboost==1.2.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
docopt==0.6.2
easyocr==1.7.1
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.4
Flask==3.0.0
flast==1.3
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2023.9.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.58.0
h5py==3.9.0
idna==3.4
imageio==2.31.5
install==1.3.5
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
labelImg==1.8.6
lazy_loader==0.3
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.16.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycountry==22.3.5
Pygments==2.16.1
pyparsing==3.1.1
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-utils==3.7.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.3
seaborn==0.12.2
shapely==2.0.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.2
sumy==0.11.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.197
urllib3==2.0.6
utils==1.0.1
wcwidth==0.2.6
Werkzeug==3.0.0
wrapt==1.15.0
